Correct build-many-glibcs.py arm-linux-gnueabihf configurations.
The conventional configure triplet for ARM GNU/Linux with hard-float
ABI is arm-*-linux-gnueabihf.  However, GCC does not automatically use
the hard-float ABI based on that triplet.  This patch fixes
build-many-glibcs.py to pass --with-float=hard so that the
arm-linux-gnueabihf configurations actually build with the intended
ABI.

Tested building the affected configurations with build-many-glibcs.py.

	* scripts/build-many-glibcs.py (Context.add_all_configs): Use
	--with-float=hard for arm-linux-gnueabihf configurations.