manual: Touch up documentation for strftime and ALTMON_* constants.
"%OB" is considered a conversion specifier ("B" is the format
specifier), and the list of format specifiers for months in the
description of the optional "O" modifier was incomplete.  A
cross-reference from the ALTMON_* constants to the strftime section
is also provided.  Lastly, some grammatical fixes (commas) are made
and paragraphs refactored (rewrapped).

	* manual/locale.texi (ALTMON_1, ALTMON_2, ALTMON_3, ALTMON_4,
	ALTMON_5, ALTMON_6, ALTMON_7, ALTMON_8, ALTMON_9, ALTMON_10,
	ALTMON_11, ALTMON_12): Improve documentation.
	* manual/time.texi (strftime): Likewise.