Fix -Os log1p, log1pf build (bug 21314).
As reported in bug 21314, building log1p and log1pf fails with -Os
because of a spurious -Wmaybe-uninitialized warning (reported there
for GCC 5 for MIPS, I see it also with GCC 7 for x86_64).  This patch,
based on the patches in the bug, fixes this using the DIAG_* macros.

Tested for x86_64 with -Os that this eliminates those warnings and so
allows the build to progress further.

2018-02-01  Carlos O'Donell  <carlos@redhat.com>
	    Ramin Seyed-Moussavi  <lordrasmus@gmail.com>
	    Joseph Myers  <joseph@codesourcery.com>

	[BZ #21314]
	* sysdeps/ieee754/dbl-64/s_log1p.c: Include <libc-diag.h>.
	(__log1p): Disable -Wmaybe-uninitialized for -Os around
	computation using c.
	* sysdeps/ieee754/flt-32/s_log1pf.c: Include <libc-diag.h>.
	(__log1pf): Disable -Wmaybe-uninitialized for -Os around
	computation using c.