Update syscall-names.list for 4.15.
This patch updates sysdeps/unix/sysv/linux/syscall-names.list for
Linux 4.15.  There only appears to be one new syscall to add to the
list.  (The riscv_flush_icache syscall is *not* added because for
whatever reason it doesn't appear in the uapi asm/unistd.h; only in
arch/riscv/include/uapi/asm/syscalls.h, which is only included by the
non-uapi asm/unistd.h - and only syscalls whose __NR_* macros are
defined in the uapi asm/unistd.h are relevant for this list.)

Tested for x86_64, and with build-many-glibcs.py.

	* sysdeps/unix/sysv/linux/syscall-names.list: Update kernel
	version to 4.15.
	(s390_sthyi): New syscall.