Use xmalloc in tst-setcontext-fpscr.c (bug 19668).
Bug 19668 reports an unchecked malloc call in the test
sysdeps/powerpc/fpu/tst-setcontext-fpscr.c.  This patch makes that
test use xmalloc.  It does not otherwise move this test to the
support/ infrastructure or support/test-driver.c; the test has various
uses of exit and _exit on error cases, and uses atexit, and while I
think those things would all still work in the context of
test-driver.c, it's not an immediately obvious conversion the way it
would be for many tests that don't use test-driver.c.

Tested for powerpc.

	[BZ #19668]
	* sysdeps/powerpc/fpu/tst-setcontext-fpscr.c: Include
	<support/support.h>.  Do not include <malloc.h>.
	(query_auxv): Use xmalloc instead of malloc.